git-svn-id: https://i-mscp.svn.sourceforge.net/svnroot/i-mscp/trunk@2455 15a38060-5cfe-4f5c-bf9d-b5d42b1378a7

+++ engine/setup/ispcp-setup-methods.pl
@@ -2417,7 +2417,10 @@ sub setup_gui_pma {
 
 		foreach($main::cfg{'PMA_USER'}, $pma_sql_user) {
 
-			$i=1 and next if($main::cfg{'PMA_USER'} eq $pma_sql_user && $i == 0);
+			if($main::cfg{'PMA_USER'} eq $pma_sql_user && $i == 0) {
+				$i++;
+				next;
+			}
 
 			$sql = "DELETE FROM tables_priv WHERE Host = '$hostname'
 				AND Db = 'mysql' AND User = '$_';";
